requirements.txt
asgiref==3.3.1
attrs==20.3.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
codecov==2.0.5
coreapi==2.3.3
coreschema==0.0.4
coverage==4.3.4

dj-rest-auth==2.1.3
Django==3.1.5

django-allauth==0.44.0
django-cors-headers==3.7.0

django-rest-auth==0.9.5
django-templated-mail==1.1.1
djangorestframework==3.12.2


iniconfig==1.1.1
itypes==1.2.0

MarkupSafe==1.1.1
oauthlib==3.1.0

pluggy==0.13.1
psycopg2==2.8.6
psycopg2-binary==2.8.6
py==1.10.0
pycparser==2.20
PyJWT==2.0.1
pyparsing==2.4.7
pytest==6.2.1
pytest-django==4.1.0
python-dotenv==0.15.0
python3-openid==3.2.0
pytz==2020.5
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0